# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "goose-server",
    "description": "An AI agent",
    "contact": {
      "name": "AAIF",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.41.0"
  },
  "paths": {
...
    "/sessions/{session_id}": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session history retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/sessions/{session_id}/extension_data": {
      "put": {
        "tags": [
          "Session Management"
        ],
        "operationId": "update_extension_data",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Extension data updated successfully"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/sessions/{session_id}/extensions": {
      "get": {
        "tags": [
          "Session Management"
        ],
        "operationId": "get_session_extensions",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session extensions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionExtensionsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/sessions/{session_id}/fork": {
      "post": {
        "tags": [
          "Session Management"
        ],
        "operationId": "fork_session",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session forked successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - truncate=true requires timestamp"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/sessions/{session_id}/name": {
      "put": {
        "tags": [
          "Session Management"
        ],
        "operationId": "update_session_name",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionNameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session name updated successfully"
          },
          "400": {
            "description": "Bad request - Name too long (max 200 characters)"
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
    "/sessions/{session_id}/user_recipe_values": {
      "put": {
        "tags": [
          "Session Management"
        ],
        "operationId": "update_session_user_recipe_values",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "description": "Unique identifier for the session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionUserRecipeValuesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Session user recipe values updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateSessionUserRecipeValuesResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Invalid or missing API key"
          },
          "404": {
            "description": "Session not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Author": {
        "type": "object",
        "properties": {
          "contact": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Conversation": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Message"
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "ExtensionData": {
        "type": "object",
        "description": "Extension data containing all extension states\nKeys are in format \"extension_name.version\" (e.g., \"todo.v0\")",
        "additionalProperties": {}
      },
...
      "ForkRequest": {
        "type": "object",
        "required": [
          "truncate",
          "copy"
        ],
        "properties": {
          "copy": {
            "type": "boolean"
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "truncate": {
            "type": "boolean"
          }
        }
      },
      "ForkResponse": {
        "type": "object",
        "required": [
          "sessionId"
        ],
        "properties": {
          "sessionId": {
            "type": "string"
          }
        }
      },
...
      "GooseMode": {
        "type": "string",
        "enum": [
          "auto",
          "approve",
          "smart_approve",
          "chat"
        ]
      },
...
      "Message": {
        "type": "object",
        "description": "A message to or from an LLM",
        "required": [
          "role",
          "created",
          "content",
          "metadata"
        ],
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageContent"
            }
          },
          "created": {
            "type": "integer",
            "format": "int64"
          },
          "id": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "$ref": "#/components/schemas/MessageMetadata"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "ModelConfig": {
        "type": "object",
        "required": [
          "model_name",
          "toolshim"
        ],
        "properties": {
          "context_limit": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "max_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "model_name": {
            "type": "string"
          },
          "reasoning": {
            "type": "boolean",
            "nullable": true
          },
          "request_params": {
            "type": "object",
            "description": "Provider-specific request parameters (e.g., anthropic_beta headers)",
            "additionalProperties": {},
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "nullable": true
          },
          "toolshim": {
            "type": "boolean"
          },
          "toolshim_model": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "Recipe": {
        "type": "object",
        "required": [
          "title",
          "description"
        ],
        "properties": {
          "activities": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "author": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Author"
              }
            ],
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionConfig"
            },
            "nullable": true
          },
          "instructions": {
            "type": "string",
            "nullable": true
          },
          "parameters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RecipeParameter"
            },
            "nullable": true
          },
          "prompt": {
            "type": "string",
            "nullable": true
          },
          "response": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Response"
              }
            ],
            "nullable": true
          },
          "retry": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RetryConfig"
              }
            ],
            "nullable": true
          },
          "settings": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ],
            "nullable": true
          },
          "sub_recipes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubRecipe"
            },
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
      "RecipeParameter": {
        "type": "object",
        "required": [
          "key",
          "input_type",
          "requirement",
          "description"
        ],
        "properties": {
          "default": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string"
          },
          "input_type": {
            "$ref": "#/components/schemas/RecipeParameterInputType"
          },
          "key": {
            "type": "string"
          },
          "options": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "requirement": {
            "$ref": "#/components/schemas/RecipeParameterRequirement"
          }
        }
      },
...
      "RecipeParameterRequirement": {
        "type": "string",
        "enum": [
          "required",
          "optional",
          "user_prompt"
        ]
      },
...
      "Response": {
        "type": "object",
        "properties": {
          "json_schema": {
            "nullable": true
          }
        }
      },
...
      "RetryConfig": {
        "type": "object",
        "description": "Configuration for retry logic in recipe execution",
        "required": [
          "max_retries",
          "checks"
        ],
        "properties": {
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SuccessCheck"
            },
            "description": "List of success checks to validate recipe completion"
          },
          "max_retries": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum number of retry attempts before giving up",
            "minimum": 0
          },
          "on_failure": {
            "type": "string",
            "description": "Optional shell command to run on failure for cleanup",
            "nullable": true
          },
          "on_failure_timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Timeout in seconds for on_failure commands (default: 600 seconds)",
            "nullable": true,
            "minimum": 0
          },
          "timeout_seconds": {
            "type": "integer",
            "format": "int64",
            "description": "Timeout in seconds for individual shell commands (default: 300 seconds)",
            "nullable": true,
            "minimum": 0
          }
        }
      },
      "Role": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "user"
            ]
          },
          {
            "type": "string",
            "enum": [
              "assistant"
            ]
          }
        ]
      },
...
      "Session": {
        "type": "object",
        "required": [
          "id",
          "working_dir",
          "name",
          "created_at",
          "updated_at",
          "extension_data",
          "message_count"
        ],
        "properties": {
          "accumulated_cost": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "accumulated_usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "archived_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "conversation": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Conversation"
              }
            ],
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "extension_data": {
            "$ref": "#/components/schemas/ExtensionData"
          },
          "goose_mode": {
            "$ref": "#/components/schemas/GooseMode"
          },
          "id": {
            "type": "string"
          },
          "last_message_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "last_message_snippet": {
            "type": "string",
            "nullable": true
          },
          "message_count": {
            "type": "integer",
            "minimum": 0
          },
          "model_config": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ModelConfig"
              }
            ],
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "project_id": {
            "type": "string",
            "nullable": true
          },
          "provider_name": {
            "type": "string",
            "nullable": true
          },
          "recipe": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Recipe"
              }
            ],
            "nullable": true
          },
          "schedule_id": {
            "type": "string",
            "nullable": true
          },
          "session_type": {
            "$ref": "#/components/schemas/SessionType"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          },
          "user_recipe_values": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "nullable": true
          },
          "user_set_name": {
            "type": "boolean"
          },
          "working_dir": {
            "type": "string"
          }
        }
      },
...
      "SessionExtensionsResponse": {
        "type": "object",
        "required": [
          "extensions"
        ],
        "properties": {
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtensionConfig"
            }
          }
        }
      },
...
      "SessionType": {
        "type": "string",
        "enum": [
          "user",
          "scheduled",
          "sub_agent",
          "hidden",
          "terminal",
          "gateway",
          "acp"
        ]
      },
...
      "Settings": {
        "type": "object",
        "properties": {
          "goose_model": {
            "type": "string",
            "nullable": true
          },
          "goose_provider": {
            "type": "string",
            "nullable": true
          },
          "max_turns": {
            "type": "integer",
            "nullable": true,
            "minimum": 0
          },
          "temperature": {
            "type": "number",
            "format": "float",
            "nullable": true
          }
        }
      },
...
      "UpdateSessionNameRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Updated name for the session (max 200 characters)"
          }
        }
      },
...
      "UpdateSessionUserRecipeValuesRequest": {
        "type": "object",
        "required": [
          "userRecipeValues"
        ],
        "properties": {
          "userRecipeValues": {
            "type": "object",
            "description": "Recipe parameter values entered by the user",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "UpdateSessionUserRecipeValuesResponse": {
        "type": "object",
        "required": [
          "recipe"
        ],
        "properties": {
          "recipe": {
            "$ref": "#/components/schemas/Recipe"
          }
        }
      },
...
      "Usage": {
        "type": "object",
        "description": "`input_tokens` is the total input including cache read/write tokens;\nthe cache fields are breakdown subsets of it. Parsers for providers\nthat report cache tokens separately from input (e.g. Anthropic,\nBedrock) must fold them into `input_tokens`.",
        "properties": {
          "cache_read_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "cache_write_input_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "input_tokens": {
            "type": "integer",
            "format": "int32",
            "description": "All prompt tokens, including any served from or written to cache.\n`cache_read_input_tokens` and `cache_write_input_tokens` are subsets of this.",
            "nullable": true
          },
          "output_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "total_tokens": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          }
        }
      },
...
    }
  }
}